Use size_t as type for buffer length in Process

+++ Process.h
@@ -176,7 +176,7 @@ void Process_printTime(RichString* str, unsigned long long totalHundredths);
 
 void Process_fillStarttimeBuffer(Process* this);
 
-void Process_outputRate(RichString* str, char* buffer, int n, double rate, int coloring);
+void Process_outputRate(RichString* str, char* buffer, size_t n, double rate, int coloring);
 
 void Process_display(const Object* cast, RichString* out);
